pyproject.toml
[project]
name = "quiz_generator_fipl_2026"
version = "0.1.0"
description = "Quiz generator for FiPL 2026"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "annotated-doc==0.0.4",
  "annotated-types==0.7.0",
  "anyio==4.13.0",
  "astroid==4.0.4",
  "blis==1.3.3",
  "catalogue==2.0.10",
  "certifi==2026.2.25",
  "charset-normalizer==3.4.6",
  "click==8.3.1",
  "cloudpathlib==0.23.0",
  "colorama==0.4.6",
  "confection==1.3.3",
  "coverage==7.13.5",
  "cymem==2.0.13",
  "dill==0.4.1",
  "filelock==3.25.2",
  "flake8==7.3.0",
  "fr_core_news_sm==3.8.0",
  "fsspec==2026.2.0",
  "gensim==4.4.0",
  "h11==0.16.0",
  "hf-xet==1.4.2",
  "httpcore==1.0.9",
  "httpx==0.28.1",
  "huggingface_hub==1.7.2",
  "idna==3.11",
  "isort==8.0.1",
  "Jinja2==3.1.6",
  "joblib==1.5.3",
  "librt==0.8.1",
  "lxml==6.0.2",
  "markdown-it-py==4.0.0",
  "MarkupSafe==3.0.3",
  "mccabe==0.7.0",
  "mdurl==0.1.2",
  "mpmath==1.3.0",
  "murmurhash==1.0.15",
  "mypy==1.19.1",
  "mypy_extensions==1.1.0",
  "networkx==3.6.1",
  "nltk==3.9.4",
  "numpy==2.4.3",
  "packaging==26.0",
  "pandas==3.0.1",
  "pathspec==1.0.4",
  "platformdirs==4.9.4",
  "preshed==3.0.13",
  "pycodestyle==2.14.0",
  "pydocstyle==6.3.0",
  "pydoctest==0.2.1",
  "pyflakes==3.4.0",
  "Pygments==2.19.2",
  "pylint==4.0.5",
  "python-dateutil==2.9.0.post0",
  "python-docx==1.2.0",
  "PyYAML==6.0.3",
  "regex==2026.2.28",
  "requests==2.32.5",
  "rich==14.3.3",
  "safetensors==0.7.0",
  "scipy==1.17.1",
  "shellingham==1.5.4",
  "six==1.17.0",
  "smart_open==7.5.1",
  "snowballstemmer==3.0.1",
  "spacy==3.8.13",
  "spacy-legacy==3.0.12",
  "spacy-loggers==1.0.5",
  "srsly==2.5.3",
  "sympy==1.14.0",
  "thinc==8.3.13",
  "tokenizers==0.22.2",
  "tomlkit==0.14.0",
  "torch==2.11.0",
  "tqdm==4.67.3",
  "transformers==5.3.0",
  "typer==0.24.1",
  "typing-inspection==0.4.2",
  "typing_extensions==4.15.0",
  "tzdata==2025.3",
  "urllib3==2.6.3",
  "wasabi==1.1.3",
  "weasel==1.0.0",
  "wrapt==2.1.2"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages]
find = {where = ["src"]}

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
py_version = "311"
skip_gitignore = true
line_length = 100
order_by_type = true
remove_redundant_aliases = true
group_by_package = true
force_alphabetary_sort_within_sections = true
src_paths = [
    "core",
    "exercises",
    "formatters",
    "generators"
]

[tool.mypy]
python_version = "3.11"
exclude = "^(venv.*|tests/|docs/)"
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
show_error_codes = true
warn_unused_ignores = false

disable_error_code = [
    "union-attr",
    "var-annotated",
    "attr-defined",
]

[[tool.mypy.overrides]]
module = [
    "datasets",
    "evaluate",
    "transformers",
    "torch.*",
    "tests.*",
    "conftest"
]
ignore_errors = true
ignore_missing_imports = true

[tool.pylint.main]
fail-under = 10.0
jobs = 1
py-version = "3.11"

[tool.pylint.format]
max-line-length = 100
exclude = "tests/"

[tool.pylint.MASTER]
ignore=["tests"]

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = ["venv*", "build", ".git", "dist", "output", "src"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.logging]
version = 1
disable_existing_loggers = true
style = "%"
datefmt = "%d/%m/%Y %H:%M:%S"
propagate = true

[tool.logging.loggers." "]
level = "INFO"
handlers = ["console"]

[tool.logging.handlers.console]
class = "logging.StreamHandler"
level = "INFO"
formatter = "formatter"

[tool.logging.formatters.formatter]
class = "logging.Formatter"
format = "%(asctime)s %(name)s: %(message)s"
